Re-render after a timeout when debouncing LSP messages
The current debounce function has a bug that can prevent rendering when
we should: if a language server sends multiple messages within an
LSP_DEADLINE we correctly debounce them but do not trigger a render when
the language server is done. For example a language server may send many
progress messages but we may miss rendering the work done progress end
message, preventing the spinner from going away.

The problem is the `editor.language_servers.incoming.is_empty()` check.
This is actually [`SelectAll::is_empty`][method] from futures_util: it
checks that the set of streams is empty rather than that there are no
pending messages in any stream.

I don't believe there is a way to check that any streams have pending
messages, so this commit uses a timer instead which triggers a render
when the debounce is done. So we render once per LSP_DEADLINE in cases
when a language server sends many messages.

[method]: https://docs.rs/futures-util/0.2.1/futures_util/stream/struct.SelectAll.html#method.is_empty